Create ord-approving-editing-of-ordinances.txt
The purpose of this ordinance is to approve the editing and 
inclusion of legislation to Springboro's Codified Ordinances passed by 
City Council between April of 2012 and March of 2013. Over the past 
year, local and state amendments involved changes to approximately 150 
pages of the City’s code, which include amendments to the Streets, 
Utilities & Public Services, General Offenses, Administrative and Planning 
& Zoning Codes as well as the City’s fee schedule for permits and 
services. In addition, this ordinance approves and adopts amendments to 
the Traffic and General Offenses Codes in accordance with enactments 
of the Ohio General Assembly. 
All of the amendments included in this code update are printed as 
the 2013 Replacement Pages to the Codified Ordinances, and properly 
format and classify the new and amended codes. The replacement 
pages are filed in the Clerk of Council’s office and copies distributed to 
City departments. Internet and desktop editions of the City's code are 
updated accordingly